Added morse IP flash and stopped double running
Added in a feature that will flash the last section of the IP address (eg. 192.168.1.155 takes just 155) in binary. This is to ensure the IP address is what you expect it to be if the VNC server is running and pupils still can't log in. It is assumed that only the last section of the IP is critical. The number will flash only a nibble if the number is less than 16, otherwise it will flash a whole byte. 

Also added in a check against a bool in a file to ensure the program only runs once, otherwise it'll mess up the morse!